Implement hard delete in UI (#131)
This implements `super-powers` only feature to hard delete all the
application data from DB.

This is meant to be used only to cleanup after testing or if we ever
receive some illegitimate garbage (spam). Otherwise we should not be
deleting data. Rejected application just get rejected. If there is we
get GDPR request we just null the data in application and remove files
containing personal information but keep the application entry as a
record.