New README.md focused on new developers experience and corresponding `falke.nix` fixes (#205)
Creates a new `README.md` while keeping the original one renamed to
`README.old.md`. The content of the new `README.md` briefly explains
individual components of the system and then presents a chronological
order of steps that are needed to bootstrap, build and run all of them.
It also discuss differences of the process on Ubuntu and Windows with
WSL running a Ubuntu. In the future, additional notes what are the key
difference for other operating systems, especially bare-metal NixOS, can
be added. Few changes to `flake.nix` were needed to enable this flow
work out-of-the box.

The new steps require a script for adding development Orca users added
by https://github.com/ictunion/main-system/pull/207

---------

[anonymized]
[anonymized]